coreth sync: 0.12.4-rc.0 -> 0.12.7 (#975)
* coreth sync: 0.12.4-rc.0 -> 0.12.7

* Review coreth 0.12.7 (#980)

* Apply suggested changes from https://github.com/ava-labs/coreth/pull/381/files

* drop reexec=0

* add txHashes

* bump avalanchego

* remove manager

* fix compat test

* bump version for e2e tests

---------

+++ eth/gasprice/feehistory_test.go
@@ -45,7 +45,7 @@ func TestFeeHistory(t *testing.T) {
 	var cases = []struct {
 		pending      bool
 		maxCallBlock uint64
-		maxBlock     int
+		maxBlock     uint64
 		count        uint64
 		last         rpc.BlockNumber
 		percent      []float64
